src/chatbot_web_demo/requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib== 0.7.5
chromadb==0.5.4
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
deepdiff==7.0.1
Deprecated==1.2.14
detectron2==0.6             
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython== 2.6.1
einops==0.8.0
email_validator==2.2.0
emoji==2.12.1
environs==9.5.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
filetype==1.2.0
FlagEmbedding== 1.2.10             
flatbuffers== 24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
humanfriendly==10.0
hydra-core== 1.3.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.9
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain== 0.2.8
langchain-community==0.2.7
langchain-core==0.2.20
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.88
layoutparser== 0.3.4
llama-cloud==0.0.6
llama-index==0.10.49
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.49
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-postprocessor-flag-embedding-reranker==0.1.3
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-chroma==0.1.10
llama-index-vector-stores-milvus==0.1.20
llama-parse==0.4.8
llamaindex-py-client==0.1.19
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
milvus-lite==2.4.8
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.23.0
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.14
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.6
overrides== 7.7.0
packaging==24.1
pandas==1.4.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
peft==0.11.1
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.17.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.5.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.4
pyparsing==3.1.2
pypdf==4.3.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
pytesseract==0.3.10
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.9
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.9.3
sentence-transformers== 3.0.1
sentencepiece==0.2.0
setuptools==70.3.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit==1.34.0
striprtf==0.0.26
sympy ==1.13.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typer==0.12.3
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata== 2024.1
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.24.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets== 12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash== 3.4.1
yacs==0.1.8
yarl== 1.9.4
zipp==3.19.2